# mnemonic/internal/http/ui/openapi.yaml
openapi: 3.0.3
info:
  title: Skillgrid Mnemonic HTTP API
  description: |-
    Persistent memory, code index, and web research cache for AI agents.

    Write routes (POST) optionally require an `Authorization: Bearer <token>`
    header when the server is started with `SKILLGRID_HTTP_TOKEN` set.
    Most routes take a `project` query parameter naming the project store
    (see `GET /projects` for available IDs).
  version: "0.1.0"
servers:
  - url: /
...
paths:
...
  /context:
    get:
      tags: [memory]
      operationId: context
      summary: Recent session context
      parameters:
        - name: project
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 5
      responses:
        "200":
          description: Recent sessions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sessions"
        "400":
          $ref: "#/components/responses/Bad"
        "500":
          $ref: "#/components/responses/Internal"
...
  /prototypes:
    get:
      tags: [prototypes]
      operationId: prototypesList
      summary: List feasibility prototypes (.skillgrid/prototypes/)
      description: |-
        Lists the directories under .skillgrid/prototypes/ and extracts a light
        summary (date, type, hypothesis, file list) from each prototype.md.
        Directories without a prototype.md (WIP) are still listed so they are
        visible rather than silently dropped. Read-only; sorted by name.
      responses:
        "200":
          description: Prototype list
          content:
            application/json:
              schema:
                type: object
                properties:
                  source:
                    type: string
                    description: Root directory scanned (relative to the served repo)
                  count: { type: integer }
                  prototypes:
                    type: array
                    items:
                      type: object
                      properties:
                        name: { type: string }
                        date: { type: string }
                        venue: { type: string, description: "The prototype.md `Type:` header" }
                        hypothesis: { type: string }
                        change:
                          type: string
                          description: Change id (specs or archive) this prototype belongs to
                        files:
                          type: array
                          items: { type: string }
              example:
                source: ".skillgrid/prototypes"
                count: 1
                prototypes:
                  - name: "001-cgo-free-vector-db"
                    date: "2026-09-24"
                    venue: "comparison (G vs viant)"
                    hypothesis: "Given a Go module, when we adopt either, then both work."
                    files: ["g.go", "prototype.md"]
        "404":
          description: No .skillgrid/prototypes/ directory
          content:
            application/json:
              schema: { type: object, properties: { error: { type: string } } }
  /prototypes/{name}/{file}:
    get:
      tags: [prototypes]
      operationId: prototypeFile
      summary: Serve one file from a feasibility prototype
      description: |-
        Serves one file under .skillgrid/prototypes/{name}/ so a generated
        index.html preview (and the relative assets it loads) can open in a
        new tab. Sandboxed to that directory: traversal or an absolute path
        → 400; a missing file or a directory → 404. Served with nosniff.
      parameters:
        - name: name
          in: path
          required: true
          description: Prototype directory name under .skillgrid/prototypes/
          schema: { type: string }
        - name: file
          in: path
          required: true
          description: File path relative to the prototype directory (e.g. index.html)
          schema: { type: string }
      responses:
        "200":
          description: The file, content type from its extension
          content:
            text/html:
              schema: { type: string }
              example: "<html></html>"
        "400":
          description: Invalid (traversal/absolute) path
          content:
            application/json:
              schema: { type: object, properties: { error: { type: string } } }
        "404":
          description: Unknown file
          content:
            application/json:
              schema: { type: object, properties: { error: { type: string } } }
  /prototype/{id...}:
    get:
      tags: [mnemonic-decisions]
      operationId: prototypeDecision
      summary: Serve one decision-companion prototype (sandboxed to .skillgrid/prototype)
      description: |-
        Serves the HTML of one visual-companion prototype, sandboxed to
        .skillgrid/prototype/ (a distinct subtree from the plural /prototypes
        above, which lists feasibility prototypes under
        .skillgrid/prototypes/). Traversal/absolute ids
        (.., leading /) → 400; unknown → 404. Served with nosniff; rendered
        client-side in a sandboxed iframe (sandbox="allow-scripts", no
        allow-same-origin) inside the decision card. Throwaway by design —
        the durable record is the decision observation + its version history.
      parameters:
        - name: id
          in: path
          required: true
          description: Prototype path relative to .skillgrid/prototype/ (e.g. <topic>/<variant>.html)
          schema: { type: string }
      responses:
        "200":
          description: Prototype HTML
          content:
            text/html:
              schema: { type: string }
              example: "<html><body><h1>Variant A</h1></body></html>"
        "400":
          description: Invalid (traversal/absolute) id
          content:
            application/json:
              schema: { type: object, properties: { error: { type: string } } }
        "404":
          description: Unknown prototype
          content:
            application/json:
              schema: { type: object, properties: { error: { type: string } } }
        "500": { $ref: "#/components/responses/Internal" }
components:
  responses:
    Bad:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/HTTPError"
    Internal:
      description: Server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/HTTPError"
...
  schemas:
...
    HTTPError:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Sessions:
      type: object
      required: [sessions]
      properties:
        sessions:
          type: array
          items:
            $ref: "#/components/schemas/Session"
...
    Session:
      type: object
      properties:
        id:
          type: string
        project:
          type: string
        directory:
          type: string
        title:
          type: string
          nullable: true
          description: Human-readable session name (optional); shown in the dashboard session list
        started_at:
          type: string
        ended_at:
          type: string
        summary:
          type: string
        status:
          type: string
...